requirements.txt
apscheduler
akshare
baostock
efinance
google-generativeai
pandas
pandas_market_calendars
psutil
pytdx
python-daemon
python-dotenv
requests
SQLAlchemy
tushare
yfinance
